chore: 디렉토리 위치 변경
