+++ src/main/java/com/kakao/cafe/repository/db/DbUserRepository.java
@@ -1,6 +1,7 @@
-package com.kakao.cafe.repository;
+package com.kakao.cafe.repository.db;
 
-import com.kakao.cafe.domain.Article;
+import com.kakao.cafe.domain.User;
+import com.kakao.cafe.repository.UserRepository;
 import org.springframework.context.annotation.Primary;
 import org.springframework.jdbc.datasource.DataSourceUtils;
 import org.springframework.stereotype.Repository;
@@ -13,105 +14,96 @@ import java.util.Optional;
 
 @Primary
 @Repository
-public class DbArticleRepository implements ArticleRepository {
+public class DbUserRepository implements UserRepository {
     private final DataSource dataSource;
 
-    public DbArticleRepository(DataSource dataSource) {
+    public DbUserRepository(DataSource dataSource) {
         this.dataSource = dataSource;
     }
 
     @Override
-    public Long save(Article article) {
-        String SQL = "INSERT INTO article (user_id, title, contents, local_date_time) VALUES (?, ?, ?, ?)";
+    public Long save(User user) {
+        String SQL = "INSERT INTO user_info (user_id, password, name, email) VALUES (?, ?, ?, ?)";
         Connection connection = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
         try {
             connection = getConnection();
             pstmt = connection.prepareStatement(SQL, Statement.RETURN_GENERATED_KEYS);
-            pstmt.setString(1, article.getUserId());
-            pstmt.setString(2, article.getTitle());
-            pstmt.setString(3, article.getContents());
-            pstmt.setTimestamp(4, Timestamp.valueOf(article.getLocalDateTime()));
+            pstmt.setString(1, user.getUserId());
+            pstmt.setString(2, user.getPassword());
+            pstmt.setString(3, user.getName());
+            pstmt.setString(4, user.getEmail());
             pstmt.executeUpdate();
 
             rs = pstmt.getGeneratedKeys();
             if (rs.next()) {
-                article.setId(rs.getLong(1));
-                return article.getId();
+                user.setId(rs.getLong(1));
+                return user.getId();
             }
         } catch (SQLException e) {
             e.printStackTrace();
-        } finally {
+        }finally {
             close(connection, pstmt, rs);
         }
         return -1L;
     }
 
     @Override
-    public Optional<Article> findById(Long id) {
-        String SQL = "SELECT id, user_id, title, contents, local_date_time FROM article WHERE id = (?)";
+    public Optional<User> findById(String userId) {
+        String SQL = "SELECT id, user_id, password, name, email FROM user_info WHERE user_id = (?)";
         Connection connection = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
         try {
             connection = getConnection();
             pstmt = connection.prepareStatement(SQL);
-            pstmt.setLong(1, id);
+            pstmt.setString(1, userId);
             rs = pstmt.executeQuery();
 
-            if (rs.next()) {
-                Article article = new Article(
-                        rs.getString("user_id"),
-                        rs.getString("title"),
-                        rs.getString("contents"),
-                        rs.getTimestamp("local_date_time").toLocalDateTime()
-                );
-                return Optional.ofNullable(article);
+            while(rs.next()){
+                User user = new User(rs.getString("user_id"), rs.getString("password"), rs.getString("name"), rs.getString("email"));
+                user.setId(rs.getLong("id"));
+                return Optional.ofNullable(user);
             }
         } catch (SQLException e) {
             e.printStackTrace();
-        } finally {
+        }finally {
             close(connection, pstmt, rs);
         }
         return Optional.empty();
     }
 
     @Override
-    public List<Article> findAll() {
-        String SQL = "SELECT id, user_id, title, contents, local_date_time FROM article";
+    public List<User> findAll() {
+        String SQL = "SELECT id, user_id, password, name, email FROM user_info";
         Connection connection = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
-        List<Article> list = new ArrayList<>();
+        List<User> list = new ArrayList<>();
 
         try {
             connection = getConnection();
             pstmt = connection.prepareStatement(SQL);
             rs = pstmt.executeQuery();
 
-            while (rs.next()) {
-                Article article = new Article(
-                        rs.getString("user_id"),
-                        rs.getString("title"),
-                        rs.getString("contents"),
-                        rs.getTimestamp("local_date_time").toLocalDateTime()
-                );
-                article.setId(rs.getLong("id"));
-                list.add(article);
+            while(rs.next()){
+                User user = new User(rs.getString("user_id"), rs.getString("password"), rs.getString("name"), rs.getString("email"));
+                user.setId(rs.getLong("id"));
+                list.add(user);
             }
             return list;
         } catch (SQLException e) {
             e.printStackTrace();
-        } finally {
+        }finally {
             close(connection, pstmt, rs);
         }
         return list;
     }
 
     @Override
-    public boolean delete(Long id) {
-        String SQL = "DELETE FROM article WHERE id = (?)";
+    public boolean delete(String userId) {
+        String SQL = "DELETE FROM user_info WHERE user_id = (?)";
         Connection connection = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -119,23 +111,23 @@ public class DbArticleRepository implements ArticleRepository {
         try {
             connection = getConnection();
             pstmt = connection.prepareStatement(SQL);
-            pstmt.setLong(1, id);
+            pstmt.setString(1, userId);
             int affectedLine = pstmt.executeUpdate();
 
-            if (affectedLine == 1) {
+            if(affectedLine == 1){
                 return true;
             }
         } catch (SQLException e) {
             e.printStackTrace();
-        } finally {
+        }finally {
             close(connection, pstmt, rs);
         }
         return false;
     }
 
     @Override
-    public void update(Long id, Article article) {
-        String SQL = "UPDATE article SET user_id = (?), title = (?), contents = (?), local_date_time = (?) WHERE id = (?)";
+    public boolean update(String userId, User updateParam) {
+        String SQL = "UPDATE user_info SET user_id = (?), password = (?), name = (?), email = (?) WHERE user_id = (?)";
         Connection connection = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -143,17 +135,22 @@ public class DbArticleRepository implements ArticleRepository {
         try {
             connection = getConnection();
             pstmt = connection.prepareStatement(SQL);
-            pstmt.setString(1, article.getUserId());
-            pstmt.setString(2, article.getTitle());
-            pstmt.setString(3, article.getContents());
-            pstmt.setTimestamp(4, Timestamp.valueOf(article.getLocalDateTime()));
-            pstmt.setLong(5, article.getId());
-            pstmt.executeUpdate();
+            pstmt.setString(1, updateParam.getUserId());
+            pstmt.setString(2, updateParam.getPassword());
+            pstmt.setString(3, updateParam.getName());
+            pstmt.setString(4, updateParam.getEmail());
+            pstmt.setString(5, updateParam.getUserId());
+            int affectedLine = pstmt.executeUpdate();
+
+            if(affectedLine == 1){
+                return true;
+            }
         } catch (SQLException e) {
             e.printStackTrace();
-        } finally {
+        }finally {
             close(connection, pstmt, rs);
         }
+        return false;
     }
 
     private Connection getConnection() {
