+++ src/main/java/com/kakao/cafe/repository/memory/MemoryArticleRepository.java
@@ -1,6 +1,7 @@
-package com.kakao.cafe.repository;
+package com.kakao.cafe.repository.memory;
 
 import com.kakao.cafe.domain.Article;
+import com.kakao.cafe.repository.ArticleRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
