+++ src/main/java/com/kakao/cafe/repository/memory/MemoryUserRepository.java
@@ -1,6 +1,7 @@
-package com.kakao.cafe.repository;
+package com.kakao.cafe.repository.memory;
 
 import com.kakao.cafe.domain.User;
+import com.kakao.cafe.repository.UserRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
