+++ src/test/java/com/kakao/cafe/repository/MemoryArticleRepositoryTest.java
@@ -1,17 +1,14 @@
 package com.kakao.cafe.repository;
 
 import com.kakao.cafe.domain.Article;
-import com.kakao.cafe.domain.User;
-import org.assertj.core.api.Assertions;
+import com.kakao.cafe.repository.memory.MemoryArticleRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class MemoryArticleRepositoryTest {
